% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adalara_no_paral.R
\name{adalara_no_paral}
\alias{adalara_no_paral}
\title{Multivariate non-parallel archetypoid algorithm for large applications (ADALARA)}
\usage{
adalara_no_paral(data, seed, N, m, numArchoid, numRep, huge, prob, type_alg = "ada", 
                 compare = FALSE, verbose = TRUE, vect_tol = c(0.95, 0.9, 0.85), 
                 alpha = 0.05, outl_degree = c("outl_strong", "outl_semi_strong", 
                 "outl_moderate"), method = "adjbox", frame)
}
\arguments{
\item{data}{Data matrix. Each row corresponds to an observation and each column 
corresponds to a variable. All variables are numeric. 
The data must have row names so that the algorithm can identify the archetypoids 
in every sample.}

\item{seed}{Integer value to set the seed. This ensures reproducibility.}

\item{N}{Number of samples.}

\item{m}{Sample size of each sample.}

\item{numArchoid}{Number of archetypes/archetypoids.}

\item{numRep}{For each \code{numArchoid}, run the archetype algorithm \code{numRep} 
times.}

\item{huge}{Penalization added to solve the convex least squares problems.}

\item{prob}{Probability with values in [0,1].}

\item{type_alg}{String. Options are 'ada' for the non-robust adalara algorithm and 
'ada_rob' for the robust adalara algorithm.}

\item{compare}{Boolean argument to compute the robust residual sum of squares 
if \code{type_alg = "ada"} and the non-robust if \code{type_alg = "ada_rob"}.}

\item{verbose}{Display progress? Default TRUE.}

\item{vect_tol}{Vector the tolerance values. Default c(0.95, 0.9, 0.85).
Needed if \code{method='toler'}.}

\item{alpha}{Significance level. Default 0.05. Needed if \code{method='toler'}.}

\item{outl_degree}{Type of outlier to identify the degree of outlierness.
Default c("outl_strong", "outl_semi_strong", "outl_moderate").
Needed if \code{method='toler'}.}

\item{method}{Method to compute the outliers. Options allowed are 'adjbox' for
using adjusted boxplots for skewed distributions, and 'toler' for using
tolerance intervals.}

\item{frame}{Boolean value to indicate whether the frame is 
computed (Mair et al., 2017) or not. The frame is made up of a subset of
extreme points, so the archetypoids are only computed on the frame. 
Low frame densities are obtained when only small portions of the data were extreme.
However, high frame densities reduce this speed-up.}
}
\value{
A list with the following elements:
\itemize{
\item cases Optimal vector of archetypoids.
\item rss Optimal residual sum of squares.
\item outliers: Outliers.
}
}
\description{
The ADALARA algorithm is based on the CLARA clustering algorithm. This is the 
non-parallel version of the algorithm. It allows to detect anomalies (outliers). 
There are two different methods to detect them: the adjusted boxplot (default 
and most reliable option) and tolerance intervals.
If needed, tolerance intervals allow to define a degree of outlierness.
}
\examples{
\dontrun{
library(Anthropometry)

# Load data:
data(mtcars)
data <- mtcars
n <- nrow(data)

# Arguments for the archetype/archetypoid algorithm:
# Number of archetypoids:
k <- 3 
numRep <- 2
huge <- 200

# Size of the random sample of observations:
m <- 10
# Number of samples:
N <- floor(1 + (n - m)/(m - k))
N
           
prob <- 0.75            

# ADALARA algorithm:
preproc <- preprocessing(data, stand = TRUE, percAccomm = 1)
data1 <- as.data.frame(preproc$data)
res_adalara <- adalara_no_paral(data1, 1, N, m, k, 
                                numRep, huge, prob, "ada_rob", FALSE, TRUE, 
                                method = "adjbox", frame = FALSE)

# Examine the results:
res_adalara

res_adalara1 <- adalara_no_paral(data1, 1, N, m, k, 
                                numRep, huge, prob, "ada_rob", FALSE, TRUE, 
                                vect_tol = c(0.95, 0.9, 0.85), 
                                alpha = 0.05, outl_degree = c("outl_strong", "outl_semi_strong", 
                                                              "outl_moderate"),
                                method = "toler", frame = FALSE)
res_adalara1                                 
                                 
}
                                                                                                                                                             
}
\references{
Eugster, M.J.A. and Leisch, F., From Spider-Man to Hero - Archetypal Analysis in 
R, 2009. \emph{Journal of Statistical Software} \bold{30(8)}, 1-23,
\url{https://doi.org/10.18637/jss.v030.i08}

Hubert, M. and Vandervieren, E., An adjusted boxplot for skewed distributions, 2008.
\emph{Computational Statistics and Data Analysis} \bold{52(12)}, 5186-5201,
\url{https://doi.org/10.1016/j.csda.2007.11.008}

Kaufman, L. and Rousseeuw, P.J., Clustering Large Data Sets, 1986.
\emph{Pattern Recognition in Practice}, 425-437.

Mair, S., Boubekki, A. and Brefeld, U., Frame-based Data Factorizations, 2017.
Proceedings of the 34th International Conference on Machine Learning, 
Sydney, Australia, 1-9.

Moliner, J. and Epifanio, I., Robust multivariate and functional archetypal analysis 
with application to financial time series analysis, 2019. 
\emph{Physica A: Statistical Mechanics and its Applications} \bold{519}, 195-208. 
\url{https://doi.org/10.1016/j.physa.2018.12.036}

Vinue, G., Anthropometry: An R Package for Analysis of Anthropometric Data, 2017.
\emph{Journal of Statistical Software} \bold{77(6)}, 1-39,
\url{https://doi.org/10.18637/jss.v077.i06}
}
\seealso{
\code{\link{do_ada}}, \code{\link{do_ada_robust}}, \code{\link{adalara}}
}
\author{
Guillermo Vinue, Irene Epifanio
}
