% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_in_r.R
\name{frame_in_r}
\alias{frame_in_r}
\title{Compute archetypes frame}
\usage{
frame_in_r(X)
}
\arguments{
\item{X}{Data frame.}
}
\value{
Vector with the observations that belong to the frame.
}
\description{
Computing the frame with the approach by Mair et al. (2017).
}
\examples{
\dontrun{
X <- mtcars
q <- frame_in_r(X)
H <- X[q,]
q
}

}
\references{
Mair, S., Boubekki, A. and Brefeld, U., Frame-based Data Factorizations, 2017.
Proceedings of the 34th International Conference on Machine Learning, 
Sydney, Australia, 1-9.
}
\author{
Sebastian Mair, code kindly provided by him.
}
