% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frobenius_norm.R
\name{frobenius_norm}
\alias{frobenius_norm}
\title{Frobenius norm}
\usage{
frobenius_norm(m)
}
\arguments{
\item{m}{Data matrix with the residuals. This matrix has 
the same dimensions as the original data matrix.}
}
\value{
Real number.
}
\description{
Computes the Frobenius norm.
}
\details{
Residuals are vectors. If there are p variables (columns),
for every observation there is a residual that there is 
a p-dimensional vector. If there are n observations, the
residuals are an n times p matrix.
}
\examples{
mat <- matrix(1:4, nrow = 2)
frobenius_norm(mat)
                 
}
\author{
Guillermo Vinue
}
