% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_prod_mat.R
\name{int_prod_mat}
\alias{int_prod_mat}
\title{Interior product between matrices}
\usage{
int_prod_mat(m)
}
\arguments{
\item{m}{Data matrix.}
}
\value{
Data matrix.
}
\description{
Helper function to compute the Frobenius norm.
}
\examples{
mat <- matrix(1:4, nrow = 2)
int_prod_mat(mat)

}
\references{
Moliner, J. and Epifanio, I., Robust multivariate and functional archetypal analysis 
with application to financial time series analysis, 2018, submitted,
\url{https://arxiv.org/abs/1810.00919}
}
\author{
Irene Epifanio
}
